backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
pymongo==4.0.1
werkzeug==2.1.2
pyjwt==2.4.0
